Moved useCalculateValues outside of component to hooks folder
